requirements.txt
asgiref==3.3.1
Django==3.1.4
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-domainname==0.1.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-smart-selects==1.5.9
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
